Set up caching for packages downloaded with Pip.
This uses the Python GitHub Action's built-in tool to cache the Pip
downloaded package directory. This cache contains the wheels (or other
package files) that were downloaded and thus avoids having to download
them again. Only the downloaded package files themselves are cached;
this does not enforce "stickiness" on what versions are installed (i.e.
if a new version of a dependency is released, that new version will be
downloaded and used; the cache will not "stick" to the old version
indefinitely).