Cache builds of libtiledbsoma in Python CI.
This caches the `build` and `dist` directories, which is the output
from the libtiledbsoma build process. It uses the hash of all files in
the libtiledbsoma directory (as well as setup.py) to ensure that when
any file is changed, the cache is invalidated and the build is done
anew.