Cache pre-commit environments. (#756)
pre-commit sets up its runtime environments in a cache directory.
If that cache does not exist, it has to download and set up the
environment anew with every run (cf. "This may take a few minutes..."
in the pre-commit output). This accounts for the majority of the runtime
of the step within GitHub actions. Caching the environments means that
this no longer needs to happen.