Create and open objects using `Thing.create()` and `.open()`.
This is the massive change to rework the whole workflow of creating and
opening SOMA objects to work more like files. `.create()` creates an
object and gives you back an open handle. `.open()` opens an object
that already exists.

This does *not* implement the `.add_whatever` part of Collections,
which should reduce a lot of test (and client!) code but I am going to
save for a future change.

Unfortunately, this kind of has to go through as one big thing because
it's a change to the semantics of the interior, and if you want the
tests to keep passing, well...

Additional notes:

- I had to sprinkle a few `@typeguard_ignore` comments around because
  TypeGuard chokes very badly on some of the generics stuff we're doing.
- Most of the tests for `whatever.exists()` and the like are gone,
  because being able to open an object means that it exists.
- Flushing could certainly be handled better. There are cases where
  the new code does it automatically, and cases where it needs to happen
  manually, and there's not too much logic as to how I've chosen to do
  one or the other apart from "maybe kind of trying to match the
  behavior to change the tests as little as possible".