Use the contents of `bld`, not `setup.py`, for binary cache.
We were (I was) using the contents of `setup.py` for the cache key when
building the tiledbsoma C++ library. This was not ideal since it meant
the cache would be busted when unrelated changes happened. We should
instead use the `bld` script, which does the actual blding, as the key.