Create `ReadWriteHandle` to manage TileDB handles. (#823)
Unlike typical file objects, when a TileDB object is opened in `w` mode,
it is *only* a writer; it can't actually read anything at all. This
introduces a class that allows you to manage both of them in one place,
and more easily flush writes and ensure that they are visible.