Make open semantics `read` XOR `write`.
Following discussion, it was clear that we would not be able to provide
a consistent and understandable model for accessing data if it was both
readable and writable. Therefore, this change separates the two apart
(for data storage; for collections, we need to be able to see what is in
them while writing).

We still need to keep around the read handle for certain activities
(metadata, groups), but its use is minimized, and the "flush" method
is explicitly marked as a hack that we should avoid using to the extent
it is possible.

In the interest of making there be exactly one source of truth for
mode and context, those are now stored *exclusively* in
`ReadWriteHandle`.

The `delete` method has been removed, since deleting the storage backing
your active object is not particularly sensible. It should be recreated
as a `tiledbsoma`-level object, since it is not related to the
particular type it is attached to.