[python] Fix `use_relative_uri` for tiledbsoma.io [post-824] (#833)
* Create and open objects using `Thing.create()` and `.open()`.

This is the massive change to rework the whole workflow of creating and
opening SOMA objects to work more like files. `.create()` creates an
object and gives you back an open handle. `.open()` opens an object
that already exists.

This does *not* implement the `.add_whatever` part of Collections,
which should reduce a lot of test (and client!) code but I am going to
save for a future change.

Unfortunately, this kind of has to go through as one big thing because
it's a change to the semantics of the interior, and if you want the
tests to keep passing, well...

Additional notes:

- I had to sprinkle a few `@typeguard_ignore` comments around because
  TypeGuard chokes very badly on some of the generics stuff we're doing.
- Most of the tests for `whatever.exists()` and the like are gone,
  because being able to open an object means that it exists.
- Flushing could certainly be handled better. There are cases where
  the new code does it automatically, and cases where it needs to happen
  manually, and there's not too much logic as to how I've chosen to do
  one or the other apart from "maybe kind of trying to match the
  behavior to change the tests as little as possible".

* Make open semantics `read` XOR `write`.

Following discussion, it was clear that we would not be able to provide
a consistent and understandable model for accessing data if it was both
readable and writable. Therefore, this change separates the two apart
(for data storage; for collections, we need to be able to see what is in
them while writing).

We still need to keep around the read handle for certain activities
(metadata, groups), but its use is minimized, and the "flush" method
is explicitly marked as a hack that we should avoid using to the extent
it is possible.

In the interest of making there be exactly one source of truth for
mode and context, those are now stored *exclusively* in
`ReadWriteHandle`.

The `delete` method has been removed, since deleting the storage backing
your active object is not particularly sensible. It should be recreated
as a `tiledbsoma`-level object, since it is not related to the
particular type it is attached to.

* Move build_tiledb_schema into common_nd_array.

* [python] Fix `use_relative_uri` for tiledbsoma.io

* CI

* lint

* CI

* ruff 0.0.226 in sandbox

* CI

* CI

* lint: ruff fixes

* lint

* CI

---------

Co-authored-by: Paul Fisher <paul.fisher@tiledb.com>
Co-authored-by: Andrew Tolopko <atolopko@chanzuckerberg.com>