Load TileDB metadata immediately upon opening the object.
As far as a reader is concerned, the TileDB metadata of a stored object
does not change over the duration of that object being open. So by
immediately loading all the metadata upon opening the TileDB object,
we have a consistent view of it and (in the future) will not need to
keep the read handle around.