Keep around only the TileDB read handle or the write handle.
This change replaces the existing `ReadWriteHandle` with a new `Handle`
type (with specializations for arrays and groups) that maintains only
EITHER the read handle OR the write handle, but not both. Information
that is needed long-term from reading is loaded immediately when the
handle is created and then cached from then on.