Refactor processing of URIs for collection creation/setting.
This separates the way URIs are processed when creating a new child
element from the way they are processed when adding an existing SOMA
object into the collection.

When creating a new child element, we already know the new relative or
absolute URI that we want to use to set it as a member, but we might
need to figure out what URI to use to create the new element.

    coll.add_new_thing("new_key", ...)
    # We want the new thing to be added at the relative URI `new_key`.
    # We need to figure out the absolute URI to create it at.
    coll.add_new_thing("abs_key", uri="file:///abs/path")
    # We want the new thing to be created at `file:///abs/path`.
    # We want to add it to the collection with that absolute URI.

Conversely, when adding an existing object into a collection, we know
the absolute URI of the collection that we are adding, but we *don't*
know the relative URI we might want to use to set the object as a member
of the collection.

    to_add = tiledbsoma.open("file:///path/to/something")
    coll["new_key"] = to_add
    # We know the absolute path of `to_add`.
    # We need to figure out whether to use a relative URI,
    # and what URI to use, depending upon our own URI.

In both cases, `tiledb://` URLs do require some special handling.