Use `**kwargs` and a function decorator to forward arguments.
For methods which primarily consist of forwarding arguments from
themselves to another backing function, this change replaces the
repeated argument list with the use of `**kwargs` to reduce repetition
and the possibility of drifting defaults. Additionally, it creates a new
`@forwards_kwargs_to` decorator to update the `__signature__` of the
wrapper function to ensure that online help still provides useful
information.

Background:

One of the primary reasons to avoid using `**kwargs` to pass arguments
directly through functions is that doing so means users no longer
receive useful help information when using a function in an interactive
environment like ipython. Instead of seeing the parameters the function
expects (and type annotations and defaults), the help text (and
parameter hints, etc.) will only show the `**kwargs`.

Conversely, using `**kwargs` is often useful and beneficial for a
library implementor (like us). Directly using `**kwargs` means that
arguments can be passed through untouched, and there is no need to
repeat argument lists and default parameters, all of which can get out
of sync from caller to callee.

These two goals are somewhat at odds. However, the standard library
provides a semi-documented workaround: a callable can provide a
`__signature__` member which is used *instead of* the actual function
signature of the called object in online help (e.g. `help(some_fn)`)
and is also used to provide parameter information in interactive shells.

How it works:

The new `@forwards_kwargs_to` decorator inspects the argument list of
both the wrapped function and the destination function and generates
a Signature which accurately reflects the combined parameters, default
values, and ordering of a function that uses `**kwargs` to directly
forward arguments to a backing function. It sets the `__signature__`
member on the wrapped function. This value is then used at runtime.

Example (from this exact code; line-wrapped for convenience)

    $ ipython
    In [1]: import tiledbsoma

    In [2]: coll = tiledbsoma.open("xxxyyy")

    In [3]: coll.add_new_dense_ndarray?
    Signature:
    coll.add_new_dense_ndarray(
        key: 'str',
        *,
        uri: 'Optional[str]' = None,
        type: pyarrow.lib.DataType,
        shape: Sequence[int],
        platform_config: Optional[Mapping[str, Any]] = None,
    ) -> 'DenseNDArray'
    Docstring:
    Adds a new DenseNDArray to this Collection.

    For details about the behavior of ``key`` and ``uri``, see
    :meth:`add_new_collection`. The remaining parameters are passed to
    the :meth:`DenseNDArray.create` method unchanged.

    In [4]: coll.add_new_dense_ndarray('something', <TAB>
    platform_config= ...
    shape=           ...
    type=            ...
    uri=             ...

Cons:

The `__signature__` value is generated at runtime and cannot be used at
static analysis time. This means that tools like `mypy` and other
static analysis–based development tools (e.g. Pylance's intellisense)
aren't able to provide detailed type checking.