Put more info in the `__repr__` of SOMA objects.
This adds more information to the general `__repr__` of SOMA objects,
including their open mode and status, and slightly reformats the
representation of collections.

It also ensures that `repr` will work on objects and collections that
have been closed, and that `repr` will only recurse into sub-elements
of a collection that have already been opened (so it doesn't eagerly
open up an entire collection tree).