Use entire dimension domain for indexing.
Updates the way NDArrays and DataFrames are indexed to use the entire
domain of a dimension rather than the non-empty domain. Since reading
past the non-empty domain is harmless, there's no need for an extra
query to figure out what it is: reading is going to return nothing.

Behavior observed at runtime is unchanged.