[python] Unify coordinate-parsing; use entire domain to slice (#936)
* Use entire dimension domain for indexing.

Updates the way NDArrays and DataFrames are indexed to use the entire
domain of a dimension rather than the non-empty domain. Since reading
past the non-empty domain is harmless, there's no need for an extra
query to figure out what it is: reading is going to return nothing.

Behavior observed at runtime is unchanged.

* Unify coordinate-parsing code into _set_reader_coords.

This pulls out all the common coordinate-parsing code and moves it into
the `_set_reader_coords` function. That function in turn has a
`_set_reader_coord` helper, which may be overridden by subclasses to
parse coordinate types specific to that subclass.

The behavior is currently unchanged.