Allow slicing with any type and with negative values. (#948)
- Removes the "slice indices must be positive" requirement,
  per discussion on https://github.com/single-cell-data/SOMA/pull/119.
- Enables slicing for string dimensions.

This change also follows up on the previous change to allow
"over-indexing" off the low end of a domain to work just as it does
when indexing off the high end of a domain: the low end is clipped to
the bottom of the domain. So for a dimension with domain [−3, 8]:

    slice(-100, 100) -> -3, 8
    slice(-5, 0) -> -3, 0
    slice(5, 10) -> 5, 8
    slice(-100, -50) -> error (too low)
    slice(50, 100) -> error (too high)

Since this uses the New And Improved `is_slice_of` function, this also
pulls in the new somacore (where that is the only change present).