[python] Use the same timestamp across read and write operations (#981)
This eliminates the distinction between the timestamps used to read and
write TileDB data. This ensures that:

- All data is written with the same timestamp.
- When one context is used to write and then read the same data,
  the same timestamp is used for both operations.
- A consistent view of a TileDB object is used for metadata when
  opening a handle for writing.

To allow users who don’t need timestamp consistency and may be running
into conflicts with the TileDB write model, we now allow specifying
None as a timestamp to opt out of consistency.