Make the default timestamp in SOMATileDBContext `None`.
Storing timestamps in the Context is problematic since it is likely that
users will want to share many parts of the Context across multiple open
calls.  This would result in every object opened on that shared context
having the same timestamp, and not just the child tree of a single open
call from user code.  This is already an issue with Cell Census, which
uses a single shared context for the entire duration of the program.

Because of this, it's better to have the timestamp be unset rather than
having it set to context creation.