Make `tiledb_timestamp` available as a datetime.
This changes the `tiledb_timestamp` member from being an integer number
of milliseconds to a `datetime.datetime`, so that users can more easily
understand when they're opening things.  The raw millisecond timestamp
is retained at `tiledb_timestamp_ms` and can be used when it exceeds
the time representable in a native datetime object.