[python] Allow all operations on a collection to share a timestamp (#1001)
This adds a `tiledb_timestamp` member to each TileDB object. This is
used when a collection member is created or opened via the collection.
Doing this means collections of objects will all be opened as of
the same timestamp (as if they had been opened eagerly upon the opening
of the outer collection).

To enable this, we add a `tiledb_timestamp` parameter to open functions.
This optional parameter defaults to the default timestamp provided by
the context. If the context's default timestamp is empty, it uses the
current date and time. This parameter can be provided as either
an integer of milliseconds or a Python datetime.

This also removes the `timestamp_start` parameter since it is unlikely
to be used in the real world and may cause confusion, particularly with
metadata (e.g., should a metadata entry set before the start_timestamp
be included? if not, that means the object cannot be opened).