[r] SOMASparseNDArray: sparseMatrix zero-based shim to facilitate soma_joinid lookup (#1313)
To load a one-based `Matrix::sparseMatrix` with the contents of a zero-based `SOMASparseNDArray`, we add one to the offsets on each dimension. However, these dimensions are usually populated with `soma_joinid` intended to match the `soma_joinid` column in the `obs` & `var` data frames, and shifting them by one makes this join operation very error-prone. 

Here we introduce a minimal shim for `sparseMatrix` that provides matrix access with zero-based indexes. To make this explicit for the user, we rename `SOMASparseNDArray$read_sparse_matrix()` to `SOMASparseNDArray$read_sparse_matrix_zero_based()`. The shim supports only minimal access operations, which is intentional to prevent "mixing" it with conventional one-based objects. If needed, the fully-featured `sparseMatrix` is recovered by calling `as.one.based()` on the shim. Thus, the default behavior is to match `soma_joinid` but an advanced user can explicitly change to one-based indexing for further use in R.

Extended discussion: #1232