[python] Make `TileDBCreateOptions` suitable for human consumption (#1408)
This updates the TileDBCreateOptions type to have attributes for each of
the settings it provides.  This means that a user can construct one
using normal-looking Python syntax, without having to build their own
JSON-like structure to be re-parsed.  TileDBCreateOptions is now
exported from tiledbsoma.options.

It also enables passing a TileDBCreateOptions directly as the
platform_config value in creation operations, implementing the pattern
that was recently legalized by the spec change.