Initialize TileDB Contexts extremely lazily.
To allow users to mess with their TileDB configs and modify them without
actually instantiating an actual TileDB Context (which entails
allocation of significant resources), we lazily wait until the caller
actually requests the Context to create it just in time. This ensures
that we don't waste resources that are not easy to reclaim (particularly
the shared global context, which is never GCable).