Log output on ctest failures
I believe this [1] failed test was ephemeral. I've had many green CI
runs [2] on my fork  this evening while attempting to debug the original
failure [1].

This PR just includes @johnkerl's branch along with setting the ctest
flag to display output on failure so that we can debug any similar
issues in the future.

[1]: https://github.com/single-cell-data/TileDB-SOMA/actions/runs/6203958202/job/16847839938
[2]: https://github.com/davisp/TileDB-SOMA/actions