Pass user config/timestamp to query condition init (#1687)
The query condition init needs the array schema and the enumeration
values. Fetching the enumeration values requires opening the array. This
was previously changed but the opening of the array did not properly
consider the the user configurations or timestamps. This change passes
the user config and timestamps to the array open in python inside the
query condition.

The overall design needs significant improvement. Opening the array
inside the query condition is likely to cause other issues as now there
are additional context created and additional performance impacts of
opening the array and closing it after enumerations are fetched.

It is left to a followup pull request to change the design to a
sustainable one that users enumerations from C++, in a similar way to the
previous passing of array schemas which avoided multiple array opens.