Fix the Python minimal CI workflow.
It now runs on `pull_request` instead of `push`, covering PRs from forks (added `workflow_dispatch` for those that might still want to run it without opening a PR), and adds a Windows matrix job (only for python 3.10 to reduce the jobs of the workflow).
Also fixed two apparent typos.