Remove arbitrary dimension limit in `used_shape`.
We were previously using `range(20)` as an arbitrarily high number of
dimensions that people were unlikely to use. Instead, `itertools.count`
provides a range-like counter that goes on forever. The function is
still guaranteed to exit because there will always be a finite number
of `soma_dim_{i}_domain_{endpoint}` metadata values.