[c++] More `SOMAGroup` refinements (#2731)
Replace contents of SOMACollection::get with a call to SOMAObject::open
Rename SOMAGroup::member_to_uri_mapping to SOMAGroup::members (this is a getter that returns a map of the child member's name and its associated SOMAGroupEntry)
Refactor SOMAGroup::members mapping to store SOMAGroupEntry values. SOMAGroupEntry stores the URI of the child member and the type of SOMAObject it is. This type info gets passed into SOMAObject::open (via SOMACollection::get) to reduce the number of unnecessary tiledb::open calls
Add an iterator for SOMACollection
Add missing documentation
Rethrow TileDBError as TileDBSOMAError where relevant