[c++,python] Fix default reads on unwritten dense arrays
* Found issue while working on spatial's `MultiscaleImage` and erroring
out when applying `read_spatial_region` to a level that had not been
written to yet
* Add new `non_empty_domain_slot_opt` alongside `non_empty_domain_slot`
that returns `std::nullopt` if it is empty rather than (0, 0)
* Fix logic in C++ `ManagedQuery::setup_read` to check if the array has
been written to yet. If it hasn't, set the subarray to use soma_dim_0's
full domain. Otherwise, use the non-empty domain as before
* Also fix logic in Python `DenseNDArray.read` to use the new
`non_empty_domain_slot_opt` and correctly set the `data_shape` to the
full domain if the array has not been written to yet
* Add unit test in pytest