Clean-up basic scene test (#3515)
The basic Scene test was hitting a sporadic failure -- likely due to a piece of the generated Scene not being fully written before it was attempted to be read back. This failure was addressed by using the `add_new_*` methods instead of creating additional arrays/groups and assigning with the `__setitem__`.