[python/cpp] Add C++ `SOMACoordinateSpace` and use it to push `Scene` metadata down to C++ (#3580)
Create a new class `SOMACoordinateSpace` in C++ and use it to read/write coordinate space metadata. Use this class to write coordinate space metadata for the `SOMAScene` class in C++. This replaces writing the metadata in the Python layer and fixes an issue where the `Scene` metadata was being written in a later timestamp than other core metadata like the SOMA datatype.