Fix undocumented reopen behavior in Python
The Python `SOMAObject.reopen` method is documented as creating a new
`SOMAObject` with the requested mode and timestamp from an existing
`SOMAObject`. The existing behavior closed and reopened the original
`SOMAObject` in its original mode in order to write metadata to the
object before reopening.

This change instead throws an error if the original object has
un-written metadata instructing the user to call close first. The
existing object is not modified.

This change allows the removal of all C++ reopen code. Simply calling
``open`` from the Python layer  is a simplier and less error prone option.