Fix undocumented reopen behavior in Python (#4106)
The Python SOMAObject.reopen method is documented as creating a new SOMAObject with the requested mode and timestamp from an existing SOMAObject. The existing behavior was to close and reopen the original SOMAObject in its original mode in order to write metadata to the object before reopening, and would then create and return a new SOMAObject with the requested mode and timestamp.

This breaking change will close the internal handle, open a new internal handle in the requested mode and timestamp, and return the original (i.e. self) from the method in the requested mode and timestamp.